# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: KACP API
  version: 0.1.0-draft
  description: |
    KCC Agent Cloud Platform 외부 API (0단계 초안).
    규칙과 내부 API는 docs/design/04-api.md 참고.
    상태 변경 요청은 X-KACP-CSRF 헤더가 필요하다.
servers:
  - url: https://app.kacp.cloud/api/v1
security:
  - session: []
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: 성공. Set-Cookie kacp_session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /auth/logout:
    post:
      tags: [auth]
      summary: 로그아웃
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      responses:
        '204': { description: 완료 }
  /auth/me:
    get:
      tags: [auth]
      summary: 내 정보
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
          description: OK
        '401': { $ref: '#/components/responses/Error' }
  /auth/password:
    post:
      tags: [auth]
      summary: 비밀번호 변경 (첫 로그인 포함)
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newPassword]
              properties:
                currentPassword: { type: string, description: 첫 로그인 설정 세션이면 생략 }
                newPassword: { type: string, minLength: 10 }
      responses:
        '204': { description: 변경됨. 다른 세션 폐기 }
        '400': { $ref: '#/components/responses/Error' }

  # ───────────── me ─────────────
...
  /me/sessions:
    get:
      tags: [me]
      summary: 내 로그인 세션
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/SessionInfo' }
    delete:
      tags: [me]
      summary: 현재 세션 외 모두 로그아웃
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      responses:
        '204': { description: 완료 }
...
  /teams/{team}/apps:
    parameters: [{ $ref: '#/components/parameters/Team' }]
    get:
      tags: [apps]
      parameters:
        - { name: work, in: query, schema: { type: string, enum: [running, sleeping, stopped, error] }, description: "sleeping = stopped + stopReason idle·limit, stopped = stopped + stopReason manual·admin" }
        - { name: public, in: query, schema: { type: string, enum: [none, publish_pending, update_pending, live] }, description: "none = 미공개(대기 요청 없음), publish_pending = 공개 요청 대기, update_pending = 공개 중 + 업데이트 요청 대기, live = 공개 중(대기 없음)" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  runningWorkCount: { type: integer }
                  runningWorkLimit: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/App' }
...
  /apps/{appId}:
    parameters: [{ $ref: '#/components/parameters/AppId' }]
    get:
      tags: [apps]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppDetail' }
    delete:
      tags: [apps]
      summary: 앱 삭제 (팀 관리자)
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      responses:
        '202': { description: "삭제 진행 (두 사본·두 데이터, 원본 폴더 유지)" }
...
  /admin/apps:
    get:
      tags: [admin]
      parameters:
        - { name: public, in: query, schema: { type: boolean } }
        - { name: team, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/App' }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: kacp_session
  parameters:
    Csrf:
      name: X-KACP-CSRF
      in: header
      required: true
      schema: { type: string }
    Team:
      name: team
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Name' }
...
    AppId:
      name: appId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [NAME_TAKEN] }
            message: { type: string, description: 사용자에게 보여줄 한국어 문구 }
            details: { type: object }
...
    Name:
      type: string
      pattern: '^[a-z0-9](?:[a-z0-9]|-(?!-)){1,28}[a-z0-9]$'
      description: 3~30자, 소문자·숫자·하이픈, -- 금지
    Space:
      type: string
      enum: [me, shared]
    PlatformRole:
      type: string
      enum: [admin, user]
...
    AppStatus:
      type: string
      enum: [starting, running, stopped, error]
    AppCopyKind:
      type: string
      enum: [work, public]
      description: work = 작업본(Private), public = 공개본(Public)
    StopReason:
      type: string
      enum: [idle, limit, manual, admin]
      description: stopped일 때만 값이 있다. 오류는 status=error + statusDetail로 표현한다
    DeployRequestKind:
      type: string
      enum: [publish, update]
    DeployRequestStatus:
      type: string
      enum: [pending, approved, rejected, cancelled]
...
    UserRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string }
        departmentName: { type: [string, 'null'], description: 동명이인 구분용 표시 }
    DepartmentRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        pathNames: { type: array, items: { type: string }, description: "최상위부터 (예: [경영지원본부, 인사팀])" }
...
    Actor:
      type: object
      description: kind=agent 이면 user 는 null ("에이전트(팀)" — OpenClaw 가 MCP 호출자 신원을 넘기지 않음, spike 05)
      properties:
        kind: { type: string, enum: [user, agent, system] }
        user: { oneOf: [{ $ref: '#/components/schemas/UserRef' }, { type: 'null' }] }
    Me:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        name: { type: string }
        platformRole: { $ref: '#/components/schemas/PlatformRole' }
        department: { oneOf: [{ $ref: '#/components/schemas/DepartmentRef' }, { type: 'null' }] }
        title: { type: [string, 'null'] }
        mustChangePassword: { type: boolean }
        csrfToken: { type: string }
...
    SessionInfo:
      type: object
      properties:
        id: { type: string }
        current: { type: boolean }
        createdAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        ip: { type: string }
        userAgent: { type: string }
...
    ResourceLimits:
      type: object
      properties:
        cpu: { type: number, description: vCPU }
        memoryMb: { type: integer }
        diskGb: { type: integer }
...
    DriveRef:
      type: object
      required: [space, path]
      properties:
        space: { $ref: '#/components/schemas/Space' }
        path: { type: string }
...
    AppCopy:
      type: object
      description: 앱의 한 사본(작업본 또는 공개본)
      properties:
        kind: { $ref: '#/components/schemas/AppCopyKind' }
        url: { type: string }
        status: { $ref: '#/components/schemas/AppStatus' }
        stopReason: { oneOf: [{ $ref: '#/components/schemas/StopReason' }, { type: 'null' }] }
        statusDetail: { type: [string, 'null'] }
        lastAccessedAt: { type: [string, 'null'], format: date-time }
        usage:
          type: object
          properties:
            cpuPct: { type: number }
            memBytes: { type: integer }
            memLimitBytes: { type: integer }
    PublicCopy:
      allOf:
        - $ref: '#/components/schemas/AppCopy'
        - type: object
          properties:
            name: { type: string }
            version: { type: integer }
            publishedAt: { type: string, format: date-time }
            approvedBy: { $ref: '#/components/schemas/UserRef' }
    App:
      type: object
      properties:
        id: { type: string, format: uuid }
        team: { type: string }
        slug: { type: string }
        creator: { $ref: '#/components/schemas/Actor' }
        work: { $ref: '#/components/schemas/AppCopy' }
        public:
          oneOf: [{ $ref: '#/components/schemas/PublicCopy' }, { type: 'null' }]
          description: null = 미공개
        pendingRequest:
          oneOf:
            - type: object
              properties:
                id: { type: string, format: uuid }
                kind: { $ref: '#/components/schemas/DeployRequestKind' }
            - type: 'null'
    AppDetail:
      allOf:
        - $ref: '#/components/schemas/App'
        - type: object
          properties:
            source: { $ref: '#/components/schemas/DriveRef' }
            runSpec:
              type: object
              properties:
                command: { type: string }
                port: { type: integer }
                runtime: { type: string, enum: [node, python, static] }
            resourceLimits: { $ref: '#/components/schemas/ResourceLimits' }
            dataUsage:
              type: object
              properties:
                workBytes: { type: integer }
                publicBytes: { type: [integer, 'null'] }
            currentRequest: { oneOf: [{ $ref: '#/components/schemas/DeployRequest' }, { type: 'null' }] }
            history:
              type: array
              items: { $ref: '#/components/schemas/DeployRequest' }
            canManage: { type: boolean, description: "삭제·공개 중지 가능 여부(팀 관리자). 시작·중지·공개 요청은 팀원 전체 가능" }
...
    DeployRequest:
      type: object
      properties:
        id: { type: string, format: uuid }
        appId: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/DeployRequestKind' }
        requestedName: { type: [string, 'null'], description: publish일 때만 }
        reason: { type: string }
        fromVersion: { type: [integer, 'null'] }
        approvedVersion: { type: [integer, 'null'] }
        diffSummary:
          oneOf:
            - type: object
              properties:
                added: { type: array, items: { type: string } }
                modified: { type: array, items: { type: string } }
                removed: { type: array, items: { type: string } }
            - type: 'null'
        status: { $ref: '#/components/schemas/DeployRequestStatus' }
        requestedBy: { $ref: '#/components/schemas/UserRef' }
        requestedAt: { type: string, format: date-time }
        decidedBy: { oneOf: [{ $ref: '#/components/schemas/UserRef' }, { type: 'null' }] }
        decidedAt: { type: [string, 'null'], format: date-time }
        decisionNote: { type: [string, 'null'] }
...